refactor #6: Google Login 메소드 로직 분리
